move RemapStorageScope to its own file
