+++ src/tir/transforms/remap_pointer_storage_scope.h
@@ -1,0 +1,44 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+/*!
+ * TODO
+ * \file remap_pointer_storage_scope.h
+ */
+#include <tvm/tir/expr.h>
+#include <tvm/tir/op.h>
+#include <tvm/tir/stmt_functor.h>
+
+#include <unordered_set>
+
+namespace tvm {
+namespace tir {
+
+class RemapStorageScope final : public StmtExprMutator {
+ public:
+  explicit RemapStorageScope(const std::unordered_map<const VarNode*, String>& new_storage_scopes);
+
+  virtual Stmt VisitStmt_(const AttrStmtNode* op);
+
+ private:
+  std::unordered_map<const VarNode*, Var> new_var_remap_;
+};
+
+}  // namespace tir
+}  // namespace tvm
